feat: extract masking routes to routes/masking.py
