+++ app.py
@@ -22213,173 +22213,6 @@ async def view_settings():
     )
 
 
-@app.route("/settings/masking", methods=["GET"])
-@require_basic_auth
-async def view_masking_settings():
-    db = get_db()
-    settings = _load_masking_settings(db)
-    return await render_template(
-        "settings_masking.html",
-        custom_keys=settings["custom_keys"],
-        custom_patterns=settings["custom_patterns"],
-        default_keys=settings["default_keys"],
-        default_patterns=settings["default_patterns"],
-        effective_key_count=len(settings["effective_keys"]),
-        effective_pattern_count=len(settings["effective_patterns"]),
-        output_masking_enabled=settings["output_masking_enabled"],
-        sql_output_masking_enabled=settings["sql_output_masking_enabled"],
-    )
-
-
-@app.route("/settings/masking/keys", methods=["POST"])
-@require_basic_auth
-async def add_masking_key():
-    db = get_db()
-    key = _masking.normalize_sensitive_key((await request.form).get("key"))
-    settings = _load_masking_settings(db)
-    if not key:
-        await flash("Sensitive key name is required", "warning")
-        return redirect(url_for("view_masking_settings"))
-    if key in settings["effective_keys"]:
-        await flash(f"Sensitive key '{key}' is already active", "info")
-        return redirect(url_for("view_masking_settings"))
-
-    custom_keys = [*settings["custom_keys"], key]
-    _save_masking_custom_keys(db, custom_keys)
-    _refresh_masking_runtime_rules(db)
-    await flash(f"Sensitive key '{key}' added", "success")
-    return redirect(url_for("view_masking_settings"))
-
-
-@app.route("/settings/masking/keys/delete", methods=["POST"])
-@require_basic_auth
-async def delete_masking_key():
-    db = get_db()
-    key = _masking.normalize_sensitive_key((await request.form).get("key"))
-    settings = _load_masking_settings(db)
-    if key not in settings["custom_keys"]:
-        await flash("Custom sensitive key not found", "warning")
-        return redirect(url_for("view_masking_settings"))
-
-    custom_keys = [item for item in settings["custom_keys"] if item != key]
-    _save_masking_custom_keys(db, custom_keys)
-    _refresh_masking_runtime_rules(db)
-    await flash(f"Sensitive key '{key}' removed", "success")
-    return redirect(url_for("view_masking_settings"))
-
-
-@app.route("/settings/masking/patterns", methods=["POST"])
-@require_basic_auth
-async def add_masking_pattern():
-    db = get_db()
-    raw_pattern = (await request.form).get("pattern")
-    settings = _load_masking_settings(db)
-    try:
-        pattern = _validate_custom_masking_pattern_for_storage(raw_pattern)
-    except (ValueError, re.error) as exc:
-        await flash(f"Invalid regex pattern: {exc}", "warning")
-        return redirect(url_for("view_masking_settings"))
-
-    if pattern in settings["effective_patterns"]:
-        await flash("That regex pattern is already active", "info")
-        return redirect(url_for("view_masking_settings"))
-
-    custom_patterns = [*settings["custom_patterns"], pattern]
-    _save_masking_custom_patterns(db, custom_patterns)
-    _refresh_masking_runtime_rules(db)
-    await flash("Custom masking pattern added", "success")
-    return redirect(url_for("view_masking_settings"))
-
-
-@app.route("/settings/masking/patterns/delete", methods=["POST"])
-@require_basic_auth
-async def delete_masking_pattern():
-    db = get_db()
-    raw_pattern = (await request.form).get("pattern")
-    settings = _load_masking_settings(db)
-    try:
-        pattern = _validate_custom_masking_pattern_for_storage(raw_pattern)
-    except (ValueError, re.error):
-        await flash("Custom masking pattern not found", "warning")
-        return redirect(url_for("view_masking_settings"))
-
-    if pattern not in settings["custom_patterns"]:
-        await flash("Custom masking pattern not found", "warning")
-        return redirect(url_for("view_masking_settings"))
-
-    custom_patterns = [item for item in settings["custom_patterns"] if item != pattern]
-    _save_masking_custom_patterns(db, custom_patterns)
-    _refresh_masking_runtime_rules(db)
-    await flash("Custom masking pattern removed", "success")
-    return redirect(url_for("view_masking_settings"))
-
-
-@app.route("/settings/masking/output", methods=["POST"])
-@require_basic_auth
-async def update_masking_output_setting():
-    db = get_db()
-    form = await request.form
-    enabled_values = form.getlist("enabled")
-    enabled = any(_is_truthy_setting(value, default=False) for value in enabled_values)
-    _set_app_setting(db, _MASKING_OUTPUT_ENABLED_SETTING, "1" if enabled else "0")
-    await flash(
-        (
-            "Global output masking enabled"
-            if enabled
-            else "Global output masking disabled across UI/JSON/notifications/GitHub issue payloads"
-        ),
-        "success",
-    )
-    return redirect(url_for("view_masking_settings"))
-
-
-@app.route("/settings/masking/sql-output", methods=["POST"])
-@require_basic_auth
-async def update_masking_sql_output_setting():
-    db = get_db()
-    form = await request.form
-    # Browser submissions can send both hidden and checkbox values for the same
-    # field name. Treat the toggle as enabled if any submitted value is truthy.
-    enabled_values = form.getlist("enabled")
-    enabled = any(_is_truthy_setting(value, default=False) for value in enabled_values)
-    _set_app_setting(db, _MASKING_SQL_OUTPUT_ENABLED_SETTING, "1" if enabled else "0")
-    await flash(
-        (
-            "SQL output masking enabled for NLQ/chart endpoints"
-            if enabled
-            else "SQL output masking disabled for NLQ/chart endpoints"
-        ),
-        "success",
-    )
-    return redirect(url_for("view_masking_settings"))
-
-
-@app.route("/api/settings/masking/preview", methods=["POST"])
-@require_basic_auth
-async def api_masking_preview():
-    payload = await request.get_json(silent=True)
-    value = (payload or {}).get("value")
-    masked = _mask_value_for_output(value) if isinstance(value, (dict, list)) else _mask_string_for_output(value)
-    return jsonify({"ok": True, "masked": masked})
-
-
-@app.route("/api/settings/masking/rules", methods=["GET"])
-@require_basic_auth
-async def api_masking_rules():
-    settings = _load_masking_settings(get_db())
-    return jsonify(
-        {
-            "ok": True,
-            "keys": settings["effective_keys"],
-            "patterns": settings["effective_patterns"],
-            "custom_keys": settings["custom_keys"],
-            "custom_patterns": settings["custom_patterns"],
-            "output_masking_enabled": settings["output_masking_enabled"],
-            "sql_output_masking_enabled": settings["sql_output_masking_enabled"],
-        }
-    )
-
-
 # ---------------------------------------------------------------------------
 # Tag Rules  GET/POST /settings/tags
 # ---------------------------------------------------------------------------
@@ -31098,12 +30931,14 @@ async def api_onboarding_create_issues():
 # already defined when the blueprint modules import from app).
 # ---------------------------------------------------------------------------
 from routes.ingest import ingest_bp as _ingest_bp  # noqa: E402
+from routes.masking import masking_bp as _masking_bp  # noqa: E402
 from routes.notifications import notifications_bp as _notifications_bp  # noqa: E402
 from routes.tags import tags_bp as _tags_bp  # noqa: E402
 from routes.apps import apps_bp as _apps_bp  # noqa: E402
 from routes.settings import settings_bp as _settings_bp  # noqa: E402
 
 app.register_blueprint(_ingest_bp)
+app.register_blueprint(_masking_bp)
 app.register_blueprint(_notifications_bp)
 app.register_blueprint(_tags_bp)
 app.register_blueprint(_apps_bp)
