+++ templates/masking_help.html
@@ -10,7 +10,7 @@
     <h4 class="fw-semibold mb-1">Output Masking Help</h4>
     <p class="text-secondary mb-0">How masking rules work across UI, JSON views, notifications, and client-side screenshot preparation.</p>
   </div>
-  <a class="btn btn-outline-secondary btn-sm" href="{{ url_for('view_masking_settings') }}" title="Back to Output Masking">
+  <a class="btn btn-outline-secondary btn-sm" href="{{ url_for('masking.view_masking_settings') }}" title="Back to Output Masking">
     <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Output Masking</span>
   </a>
 </div>
