+++ templates/settings_help.html
@@ -84,7 +84,7 @@
                 </div>
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-danger w-100"
-                       href="{{ url_for('view_masking_settings') }}">Output Masking</a>
+                       href="{{ url_for('masking.view_masking_settings') }}">Output Masking</a>
                 </div>
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-secondary w-100"
